chore: move `cargo_near::commands::abi_command::abi::run` to `cargo_near_build::build_abi`
