+++ cargo-near-build/src/lib.rs
@@ -11,3 +11,6 @@ pub mod types;
 
 // TODO: remove these exports, replace with `pub(crate)` visibility
 pub use cargo_native::{ArtifactType, DYLIB, WASM};
+
+pub use near::abi::build as build_abi;
+pub use types::near::abi::Opts as AbiOpts;
